optimized experience type = long -> int

+++ src/main/java/com/gianca1994/aowebbackend/resources/user/User.java
@@ -75,9 +75,9 @@ public class User {
     @Column
     private short level;
     @Column
-    private long experience;
+    private int experience;
     @Column
-    private long experienceToNextLevel;
+    private int experienceToNextLevel;
     @Column
     private long gold;
     @Column
